+++ ChangeLog.md
@@ -8,7 +8,7 @@
   * New-AzureRmWebAppSSLBinding
   * Get-AzureRmWebAppSSLBinding
   * Remove-AzureRmWebAppSSLBinding
-* Azure Websites: Added AseName and AseResourceGroupName parameters in New-AzureRmAppServicePlan cmdlet
+* Azure Websites: Added AseName and AseResourceGroupName parameters in New-AzureRmWebApp and New-AzureRmAppServicePlan cmdlet
   
 ## 2015.12.14 version 1.0.2
 * Azure Compute (ARM):